updated to add iterables to Vec3 / 4

+++ src/nccapy/Math/Mat4.py
@@ -43,8 +43,7 @@ class Mat4:
 
     def get_numpy(self):
         "return matrix as a numpy array ideal for WebGPU etc"
-        return np.array(self.get_matrix()).reshape([4, 4])
-
+        return np.array(self.get_matrix(), dtype=np.float32).reshape([4, 4])
     @classmethod
     def identity(cls):
         "class method to return a new identity matrix"
